add shift route, test via Postman, routes work

+++ routes/employee.js
@@ -1,15 +1,17 @@
+/* eslint-disable no-underscore-dangle */
 const express = require('express');
+const mongoose = require('mongoose');
 // const createError = require('http-errors');
 
 const router = express.Router();
 const User = require('../models/User');
 const Shift = require('../models/Shift'); // populate
-const WorkingDay = require('../models/WorkingDay');
+const WorkingDay = require('../models/WorkingDay'); // populate
 // const { checkIfLoggedIn, checkUsernameNotEmpty } = require("../middlewares");
 
 router.get('/', async (req, res, next) => {
   try {
-    const employees = await User.find({ role: 'employee' }).populate('shifts').populate('day');
+    const employees = await User.find({ role: 'employee' }).populate('shifts');
     res.json(employees);
   } catch (error) {
     next(error);
@@ -33,18 +35,23 @@ router.get('/:employeeId', async (req, res, next) => {
 // admin can change: role (to admin eg; if deletion is wanted - via delete better), shifts
 
 router.put('/:employeeId/update', async (req, res, next) => {
-  const { employeeId } = req.params;
-  const {
-    email, username, role, firstName, familyName,
-  } = req.body;
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(400).json({ message: 'Specified id is not valid' });
+      return;
+    }
+    const { employeeId } = req.params;
+    const {
+      email, username, role, firstName, familyName, shifts,
+    } = req.body;
     const employee = await User.findByIdAndUpdate(employeeId, {
       email,
       username,
       role,
       firstName,
       familyName,
-    }, { new: true }).populate('shifts');
+      shifts,
+    }, { new: true }); /* populate('shifts') */
     res.json(employee);
   } catch (error) {
     next(error);
@@ -54,10 +61,15 @@ router.put('/:employeeId/update', async (req, res, next) => {
 // deletes not only "employee", but the user himself
 
 router.delete('/:employeeId/delete', async (req, res, next) => {
-  const { employeeId } = req.params;
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(400).json({ message: 'Specified id is not valid' });
+      return;
+    }
+    const { employeeId } = req.params;
     const employee = await User.findByIdAndDelete(employeeId).populate('shifts');
-    // delete as well all shifts
+    await Shift.findByIdAndUpdate(employee.shift._id, { $pull: { employee: employee._id } });
+    await WorkingDay.findByIdAndUpdate(employee.workingDay._id, { $pull: { employeesTeam: employee._id } });
     res.json(employee);
   } catch (error) {
     next(error);
