+++ routes/user.js
@@ -1,11 +1,11 @@
 const express = require('express');
+const mongoose = require('mongoose');
 // const createError = require('http-errors');
 
 const router = express.Router();
 const User = require('../models/User');
 const Shift = require('../models/Shift'); // populate
-const WorkingDay = require('../models/WorkingDay');
-
+const WorkingDay = require('../models/WorkingDay'); // populate
 // const { checkIfLoggedIn, checkUsernameNotEmpty } = require("../middlewares");
 
 router.get('/', async (req, res, next) => {
@@ -48,18 +48,23 @@ router.post('/add', async (req, res, next) => {
   }
 });
 
-// admin has access to all registered users and can change only their role and by that add shifts
+// admin has access to all registered users and can change only their role (and by that add shifts)
+// OR maybe only change the role
 // the user himself can only change his: email, username, firstName, familyName
 // this user after his role was changed, shows up in "employees"
 
 router.put('/:userId/update', async (req, res, next) => {
-  const { userId } = req.params;
-  const {
-    email, username, role, firstName, familyName, shifts,
-  } = req.body;
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(400).json({ message: 'Specified id is not valid' });
+      return;
+    }
+    const { userId } = req.params;
+    const {
+      email, username, role, firstName, familyName,
+    } = req.body;
     const user = await User.findByIdAndUpdate(userId, {
-      email, username, role, firstName, familyName, shifts,
+      email, username, role, firstName, familyName,
     }, { new: true }).populate('shifts');
     res.json(user);
   } catch (error) {
@@ -68,8 +73,12 @@ router.put('/:userId/update', async (req, res, next) => {
 });
 
 router.delete('/:userId/delete', async (req, res, next) => {
-  const { userId } = req.params;
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      res.status(400).json({ message: 'Specified id is not valid' });
+      return;
+    }
+    const { userId } = req.params;
     const user = await User.findByIdAndDelete(userId);
     res.json(user);
   } catch (error) {
